+++ src/components/buttons/VcsFormButton.vue
@@ -115,6 +115,7 @@
    * @vue-prop {string}                                 tooltip - Text content of a tooltip which appears on hover with default delay.
    * @vue-prop {('bottom' | 'left' | 'top' | 'right')}  tooltipPosition - Position of the tooltip.
    * @vue-prop {Object<string, any>}                    tooltipProps - Properties to be passed to VcsTooltip {@link https://vuetifyjs.com/en/api/v-tooltip/#props|vuetify v-tooltip}
+   * @vue-prop {string[]}                               [customClasses] - CSS classes to customize style
    * @vue-computed {Object<string, string>}             classes - css classes applied to button, depending on variant
    * @vue-computed {boolean}                            hasDefaultSlot
    * @vue-event {MouseEvent}                            click - Emits click event when the button is clicked.
@@ -162,10 +163,18 @@
         type: Object,
         default: () => ({}),
       },
+      customClasses: {
+        type: Array,
+        default: () => [],
+      },
     },
     computed: {
       classes() {
         return {
+          ...this.customClasses.reduce(
+            (acc, cur) => ({ ...acc, [cur]: true }),
+            {},
+          ),
           'vcs-form-button--outlined': this.variant === 'outlined',
           'vcs-form-button--filled': this.variant === 'filled',
         };